starter_code/requirements.txt
# Core dependencies
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Async support
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0
websockets>=12.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
rich>=13.0.0  # For CLI output formatting
